Quote service arguments in starter command
Wrap service arguments in the service starter command with single-quotes
to fix issues when special characters are present in the arguments. For
example, if an app's Shizuku service is a nested class, starting the
service fails because "$" appears in the class name and gets resolved to
a non-existent shell veriable.

While this is only strictly needed for the class name (and potentially
the nice name, if the process name suffix includes special characters),
it doesn't hurt to quote all the strings to avoid potential problems in
the future.